refactor: use correct imports and remove old code
