+++ src-old/client/Pronote.ts
@@ -1,75 +1,32 @@
 import type { NextAuthenticationCredentials } from "~/authenticate/types";
 import type { ApiLoginInformations } from "~/api/login/informations/types";
 import type { PawnoteFetcher } from "~/utils/fetcher";
-
-import {
-  callApiUserHomework,
-  callApiUserHomeworkStatus,
-  callApiUserTimetable,
-
-  type ApiUserData
-} from "~/api";
-
-import { StudentHomework } from "~/parser/homework";
-import { Period, readOngletPeriods, OngletPeriods } from "~/parser/period";
-
 import { Session } from "~/session";
 import Queue from "~/utils/queue";
 
-import { readPronoteApiDate, transformDateToPronoteString, translateToPronoteWeekNumber } from "~/pronote/dates";
-import { getUTCDate } from "~/utils/dates";
 import { callApiUserGrades } from "~/api/user/grades";
 import { StudentGrade } from "~/parser/grade";
 import { StudentAverage } from "~/parser/average";
 import { readPronoteApiGrade } from "~/pronote/grades";
 import { callApiUserEvaluations } from "~/api/user/evaluations";
 import { StudentEvaluation } from "~/parser/evaluation";
-import { PronoteApiAccountId } from "~/constants/accounts";
-import { StudentPersonalInformation } from "~/parser/personalInformation";
-import { callApiUserPersonalInformation } from "~/api/user/personalInformation";
-import { StudentAttachment } from "~/parser/attachment";
-import { callApiUserPresence } from "~/api/user/presence";
-import { callApiUserResources } from "~/api/user/resources";
-import { callApiUserLessonResource } from "~/api/user/lessonResource";
-import { callApiUserLessonHomework } from "~/api/user/lessonHomework";
-import { StudentLessonResource } from "~/parser/lessonResource";
-import { callApiUserNews } from "~/api/user/news";
-import { StudentNews, StudentNewsItemQuestion } from "~/parser/news";
 import { callApiUserDiscussions } from "~/api/user/discussions";
 import { StudentDiscussion, StudentDiscussionsOverview } from "~/parser/discussion";
 import { PronoteApiMessagesButtonType } from "~/constants/messages";
 import { callApiUserMessages } from "~/api/user/messages";
 import { MessagesOverview } from "~/parser/messages";
-import { PronoteApiOnglets } from "~/constants/onglets";
-import { callApiUserAttendance } from "~/api/user/attendance";
-import { StudentAbsence, StudentDelay, StudentObservation, StudentPrecautionaryMeasure, StudentPunishment } from "~/parser/attendance";
 import { callApiUserMessageRecipients } from "~/api/user/messageRecipients";
 import { DiscussionCreationRecipient, FetchedMessageRecipient } from "~/parser/recipient";
-import Holiday from "~/parser/holiday";
-import type { PronoteApiNewsPublicSelf } from "~/constants/news";
-import { callApiUserNewsStatus } from "~/api/user/newsStatus";
-import Authorizations from "~/parser/authorizations";
 import type { PronoteApiUserResourceType } from "~/constants/users";
 import { callApiUserCreateDiscussionRecipients } from "~/api/user/createDiscussionRecipients";
 import { PronoteApiResourceType } from "~/constants/resources";
 import { callApiUserCreateDiscussion } from "~/api/user/createDiscussion";
 import { callApiUserCreateDiscussionMessage } from "~/api/user/createDiscussionMessage";
 import { getPronoteMessageButtonType } from "~/pronote/messages";
-import { createPronoteUploadCall } from "~/pronote/requestUpload";
-import { PronoteApiFunctions } from "~/constants/functions";
-import { callApiUserHomeworkUpload } from "~/api/user/homeworkUpload";
-import { callApiUserHomeworkRemoveUpload } from "~/api/user/homeworkRemoveUpload";
-import { callApiUserHomepage } from "~/api/user/homepage";
-import { Partner } from "~/parser/partner";
-import { callApiUserPartnerURL } from "~/api/user/partnerURL";
-import { PronoteApiDomainFrequencyType, PronoteApiMaxDomainCycle } from "~/constants/domain";
-import { parseSelection } from "~/pronote/select";
-import { TimetableOverview } from "~/parser/timetable";
+
 import { callApiUserDiscussionCommand } from "~/api/user/discussionCommand";
 import { ApiUserDiscussionAvailableCommands } from "~/api/user/discussionCommand/types";
-import type { PawnoteSupportedFormDataFile } from "~/utils/file";
 import { callApiUserGeneratePDF } from "~/api/user/generatePDF";
-import { Onglets } from "~/parser/onglets";
 
 export default class Pronote {
   private queue: Queue;
@@ -77,7 +34,6 @@ export default class Pronote {
   constructor (
     public fetcher: PawnoteFetcher,
     private session: Session,
-    credentials: NextAuthenticationCredentials,
 
     // Accessor for raw data returned from Pronote's API.
     private user: ApiUserData["output"]["data"]["donnees"],
@@ -198,49 +154,6 @@ export default class Pronote {
       });
 
       return data.donnees.listeDest.V.map((dest) => new FetchedMessageRecipient(dest));
-    });
-  }
-
-  #throwIfNotAllowedRecipientType (type: PronoteApiUserResourceType): void {
-    switch (type) {
-      case PronoteApiResourceType.Teacher:
-        if (!this.authorizations.canDiscussWithTeachers)
-          throw new Error("You can't discuss with teachers.");
-        break;
-      case PronoteApiResourceType.Student:
-        if (!this.authorizations.canDiscussWithStudents)
-          throw new Error("You can't discuss with students.");
-        break;
-      case PronoteApiResourceType.Personal:
-        if (!this.authorizations.canDiscussWithStaff)
-          throw new Error("You can't discuss with staff.");
-        break;
-    }
-  }
-
-  /**
-   * Returns a list of possible recipients when creating a discussion.
-   *
-   * This step is required before creating a discussion.
-   * It allows to know who can be the recipient of the discussion.
-   */
-  public async getRecipientsForDiscussionCreation (type: PronoteApiUserResourceType): Promise<DiscussionCreationRecipient[]> {
-    return this.queue.push(async () => {
-      this.#throwIfNotAllowedRecipientType(type);
-
-      const response = await callApiUserCreateDiscussionRecipients(this.fetcher, {
-        recipientType: type,
-        session: this.session,
-        user: {
-          type: this.user.ressource.G,
-          name: this.user.ressource.L,
-          id: this.user.ressource.N
-        }
-      });
-
-      return response.data.donnees.listeRessourcesPourCommunication.V
-        .filter((recipient) => recipient.avecDiscussion)
-        .map((r) => new DiscussionCreationRecipient(r));
     });
   }
