+++ src/api/index.ts
@@ -16,6 +16,7 @@ export * from "./assignments";
 export * from "./geolocation";
 export * from "./homepage";
 export * from "./instance";
+export * from "./new-discussion-recipients";
 export * from "./news-delete";
 export * from "./news-information-acknowledge";
 export * from "./news-read";
