+++ src/decoders/news-survey.ts
@@ -1,6 +1,6 @@
 import type { NewsCategory, SessionHandle, NewsSurvey } from "~/models";
-import { decodeNewsItem } from "./news-item";
 import { decodeNewsQuestion } from "./news-question";
+import { decodeNewsItem } from "./news-item";
 
 export const decodeNewsSurvey = (survey: any, session: SessionHandle, categories: NewsCategory[]): NewsSurvey => {
   return {
