+++ src/decoders/partner.ts
@@ -1,2 +1,2 @@
-import { Partner } from "~/models";
+import type { Partner } from "~/models";
 export const decodePartner = (partner: any): Partner => ({ sso: partner.SSO });
